# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MeisterProPR.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Reviews/{jobId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get the status and result of a review job.",
        "parameters": [
          {
            "name": "X-Ado-Token",
            "in": "header",
            "description": "ADO personal access token used solely to verify the requesting user is an authenticated ADO organisation member.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "in": "path",
            "description": "The job identifier returned from POST /reviews.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status and, once completed, its result.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewStatusResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewStatusResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing client key, or invalid ADO token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Job not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "List all review jobs for the current client.",
        "parameters": [
          {
            "name": "X-Ado-Token",
            "in": "header",
            "description": "ADO personal access token used solely to verify the requesting user is an authenticated ADO organisation member.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of review jobs, newest first.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewListItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing client key, or invalid ADO token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Submit a pull request for AI review.",
        "parameters": [
          {
            "name": "X-Ado-Token",
            "in": "header",
            "description": "ADO personal access token used solely to verify the requesting user is an authenticated ADO organisation member.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The PR details to review.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Review job accepted. Poll the returned jobId for status.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewJobResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewJobResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewJobResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing client key, or invalid ADO token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CommentSeverity": {
        "enum": [
          "info",
          "warning",
          "error",
          "suggestion"
        ],
        "type": "string",
        "description": "Severity level of a review comment."
      },
...
      "JobStatus": {
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "type": "string",
        "description": "Status of a review job."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "ReviewCommentDto": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "nullable": true
          },
          "lineNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "severity": {
            "$ref": "#/components/schemas/CommentSeverity"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO for a single review comment."
      },
      "ReviewJobResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Response returned when a review job is accepted."
      },
      "ReviewListItem": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "organizationUrl": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "repositoryId": {
            "type": "string",
            "nullable": true
          },
          "pullRequestId": {
            "type": "integer",
            "format": "int32"
          },
          "iterationId": {
            "type": "integer",
            "format": "int32"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "List item for a review job."
      },
      "ReviewRequest": {
        "type": "object",
        "properties": {
          "organizationUrl": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "repositoryId": {
            "type": "string",
            "nullable": true
          },
          "pullRequestId": {
            "type": "integer",
            "format": "int32"
          },
          "iterationId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Request payload to submit a pull request for review."
      },
      "ReviewResultDto": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string",
            "nullable": true
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewCommentDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO representing the textual review result and comments."
      },
      "ReviewStatusResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "organizationUrl": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "repositoryId": {
            "type": "string",
            "nullable": true
          },
          "pullRequestId": {
            "type": "integer",
            "format": "int32"
          },
          "iterationId": {
            "type": "integer",
            "format": "int32"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/ReviewResultDto"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Detailed status response for a review job."
      }
    }
  },
...
}